+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -5,9 +5,9 @@ import java.util.stream.Collectors;
 import kitchenpos.application.dto.TableGroupCreateRequest;
 import kitchenpos.application.dto.TableGroupCreateRequest.OrderTableRequest;
 import kitchenpos.application.dto.TableGroupResponse;
-import kitchenpos.dao.OrderTableRepository;
+import kitchenpos.order.domain.OrderTableRepository;
 import kitchenpos.dao.TableGroupRepository;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
