+++ src/main/java/kitchenpos/application/dto/TableGroupResponse.java
@@ -4,6 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.order.application.dto.OrderTableResponse;
 
 public class TableGroupResponse {
