+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,21 +1,18 @@
-package kitchenpos.application;
+package kitchenpos.order.application;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.application.dto.OrderCreateRequest;
-import kitchenpos.application.dto.OrderCreateRequest.OrderLineRequest;
-import kitchenpos.application.dto.OrderLineItemResponse;
-import kitchenpos.application.dto.OrderResponse;
-import kitchenpos.application.dto.OrderStatusChangeRequest;
-import kitchenpos.dao.OrderLineItemRepository;
-import kitchenpos.dao.OrderRepository;
-import kitchenpos.dao.OrderTableRepository;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.menu.application.dto.MenuResponse;
+import kitchenpos.order.application.dto.OrderCreateRequest;
+import kitchenpos.order.application.dto.OrderCreateRequest.OrderLineRequest;
+import kitchenpos.order.application.dto.OrderResponse;
+import kitchenpos.order.application.dto.OrderStatusChangeRequest;
+import kitchenpos.order.domain.OrderLineItemRepository;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
 import org.springframework.stereotype.Service;
