+++ src/main/java/kitchenpos/order/domain/OrderLineItemRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.order.domain;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
