+++ src/main/java/kitchenpos/order/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,6 +9,8 @@ import javax.persistence.ForeignKey;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
+import kitchenpos.domain.BaseEntity;
+import kitchenpos.domain.TableGroup;
 
 @Entity
 public class OrderTable extends BaseEntity {
