+++ src/test/java/kitchenpos/application/OrderTableServiceTest.java
@@ -3,13 +3,13 @@ package kitchenpos.application;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import kitchenpos.application.dto.OrderTableChangeEmptyRequest;
-import kitchenpos.application.dto.OrderTableChangeNumberOfGuestsRequest;
-import kitchenpos.application.dto.OrderTableCreateRequest;
-import kitchenpos.application.dto.OrderTableResponse;
-import kitchenpos.application.dto.TableGroupCreateRequest.OrderTableRequest;
-import kitchenpos.dao.OrderTableRepository;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.application.dto.OrderTableChangeEmptyRequest;
+import kitchenpos.order.application.dto.OrderTableChangeNumberOfGuestsRequest;
+import kitchenpos.order.application.dto.OrderTableCreateRequest;
+import kitchenpos.order.application.dto.OrderTableResponse;
+import kitchenpos.order.application.OrderTableService;
+import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderTable;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
